enh: remove unused code
